requirements.txt
django==1.8.13
djangorestframework==3.3.3
datadog
